Io: bound guest-supplied USB control/interrupt buffers
- LogitechG27.cpp: the change-device-mode interrupt-OUT path checked
  `buf_size < 7` but then read buf[8], and (when buf[8] == 0xf8) buf[9]
  and buf[10] — an OOB read on the 7-/8-byte boundary. Gate the
  back-to-back branch on `buf_size > 8`, and bail out when the full
  11-byte form isn't present.
- GunCon3.cpp: `(x_pos * USHRT_MAX) / x_max` ran in signed 32-bit;
  a high-DPI mouse x_pos > 32767 (common on 4K displays) overflowed
  to UB. Promote the multiplication to s64 before the divide.
- usio.cpp: the 0xFB card-reader command only checked the self-
  reported `payload_length` before dereferencing payload[3] (= buf[9])
  — guest could pass a 6-byte buffer whose checksum byte said
  `payload_length >= 5` and trigger an OOB read. Validate the real
  buf.size() against 6 + payload_length.
- Skylander.cpp: the 0x21/0x09 SET_REPORT branch read `buf[0]`
  unconditionally inside the dispatch switch; the per-case
  `ensure(buf_size == ...)` was after the access. A zero-length
  control transfer therefore OOB-read. Bail out when buf_size < 1.

+++ rpcs3/Emu/Io/GunCon3.cpp
@@ -286,9 +286,9 @@ void usb_device_guncon3::interrupt_transfer(u32 buf_size, u8* buf, u32 endpoint,
 			return;
 		}
 
-		// Expand 0..+wh to -32767..+32767
-		gc.gun_x = (mouse_data.x_pos * USHRT_MAX / mouse_data.x_max) - SHRT_MAX;
-		gc.gun_y = (mouse_data.y_pos * -USHRT_MAX / mouse_data.y_max) + SHRT_MAX;
+		// Expand 0..+wh to -32767..+32767 (force 64-bit math; x_pos*USHRT_MAX overflows s32 for high-DPI inputs)
+		gc.gun_x = static_cast<s32>(static_cast<s64>(mouse_data.x_pos) * USHRT_MAX / mouse_data.x_max - SHRT_MAX);
+		gc.gun_y = static_cast<s32>(static_cast<s64>(mouse_data.y_pos) * -USHRT_MAX / mouse_data.y_max + SHRT_MAX);
 	}
 
 	guncon3_encode(&gc, buf, m_key.data());
